pyproject.toml
[project]
name = "deltagen-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.3.18",
    "langchain>=0.3.19",
    "markitdown>=0.0.1a4",
    "unstructured>=0.16.23",
    "pillow>=11.1.0",
    "lxml>=5.3.1",
    "pi-heif>=0.21.0",
    "unstructured-inference>=0.8.7",
    "pdf2image>=1.17.0",
    "unstructured-pytesseract>=0.3.13",
    "langchain-openai>=0.3.6",
    "python-dotenv>=1.0.1",
    "chromadb>=0.6.3",
    "langgraph>=0.2.74",
    "streamlit>=1.42.2",
    "asyncio>=3.4.3",
    "loguru>=0.7.3",
    "langchain-chroma>=0.2.2",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pytest>=8.3.4",
    "ruff>=0.9.6",
]
